Fix bug with selecting color on touch up not applying
Sometimes `_pendingPickedColor` can be `null` at the point of touch up which results in no new color being set. This due to `_pendingPickedColor` being changed between the color selection and the touch up event. This commit introduces `_touchUpPendingColor` which is used exclusively to prepare and select colors for a touch up action.